fix Signed-off-by: Yi Cheng <74173148+iycheng@users.noreply.github.com>

+++ src/ray/common/ray_config_def.h
@@ -312,7 +312,7 @@ RAY_CONFIG(int64_t, redis_db_connect_retries, 600)
 RAY_CONFIG(int64_t, redis_db_connect_wait_milliseconds, 100)
 
 /// Number of retries for a redis request failure.
-RAY_CONFIG(int64_t, num_redis_request_retries, 5)
+RAY_CONFIG(size_t, num_redis_request_retries, 5)
 
 /// Exponential backoff setup. By default:
 /// 100ms, 200ms, 400ms, 800ms, 1s, 1s,...
